PBD: Always discard containers and use common offer
The BinaryDeque.offer API states that the container will be discarded if
an exception is thrown. This was not true if exceptions were thrown
prior to the attempt to write the entry. Update the implementation so
that this is now guaranteed.

Move the common code in PBDRegularSegment.offer into a common offer
which takes a BBContainer.