ENG-19497: Backport: Guarantee that catalog update happens after quiesce
[ backport cd496ad07c2584dd51a92f846b9152f1eb9190dc ]

Catalog update is an MP procedure which occurs concurrently on all
hosts. The order of operations is all sites call quiesce and then
RealVoltDB.updateCatalog(). The first site on each host to call
updateCatalog() performs the actual work. This can be an issue because
the other sites on the host might still be performing the quiesce. When
this happens the schema can be updated in export and DR before the
buffers with the data from the previous schema are processed.

This results in the new schema being attached to the old rows.