ENG-19371: Do not use memcpy to copy hidden columns
Only use memcpy to copy hidden columns if the schemas are the same
instance. If they are not the same instance copy or initialize the
hidden columns in the destination tuple by hidden column type.

This is needed because the planner and the java layer are mostly unaware
of hidden columns so a plan can be constructed which uses copy but the
hidden columns are not the same for the source and destination.