ENG-18529: Perform JAVA_GET_DRID_TRACKER as an MP
When a host becomes a dr leader it quaries all partitions for their
current trackers and merges them together. However this can race with a
RUN_EVERYWHERE_LSP which is enqueded in the MpI. It is possible for some
of the SPs to arrive before and some to arrive after the fragments of
the run everywhere. By makding the task an MP task it is guaranteed to
run after all in progress procedures.