ENG-19677: Use TxnId for replicated tables
In export and migration use the TxnId for replicated tables and the
SpHandle for partitioned tables. This allows the transaction ID column
of replication to always be in the same number space and increase in
value. This also keeps it so that the same transaction ID is used to
identify migrating rows in replicated tables.

Create a new abstract table class to hold some of the common methods and
functions between streamed and persistent tables. This includes the
method to determine which transaction ID to be used for the table.