ee: Remove unused fallible argument
All callers of PersistentTable::truncateTable() either used the default
true for fallible or passed in true. Since it was essentially unused
it can be removed. This also means that drLogTruncate and
deleteAllTuples did not use fallible either.

Also merge deleteAllTempTuples and deleteAllTuples into deleteAllTuples
without any arguments since the other argument to deleteAllTuples was
always ignored.