[Careless] optimize away some deep copier/finalize operations
on the new/different memory model:

1. Previously, both src/dst of moved tuples in batch removal are
captured in hook memory. Using the new memory model, we can safely only
capture the dst tuple (that gets overwritten) in hook.

2. Previously, inside the callback of batch removal, index is always
updated for the movement. Now, since we are guaranteed that source tuple
(the memory that gets copied to somewhere else) remains visible and
intact only in snapshot region, we *replace* index on the two tuples
only when the table is not frozen; when it is frozen, we simply add the
dst tuple to index, therefore keeping the source tuple valid in the
index. This part is most suspicious to me: should I add index to the
tuple that hook shallow copies (see #3)?

3. Previously, when adding hook memory entry for deletes or updates, we
always use deep copier when applicable. Now, we only invoke deep copier
on update: we use shallow memcpy for deleted tuples, and let batch
removal call back discriminate the frozen case, when it does not
finalize (i.e. call `decreaseStringMemCount()' and
`freeObjectColumns()') if frozen.