Fix: in batch removal, we can
no longer safely finalize all tuples that gets removed (i.e. those not being
compacted to holes) at frozen time, since we spared copy (and deep copy)
into hook memory. To correctly do this, we need to differentiate the
removed tuples, checking if it lies inside frozen boundary (in which case,
we cannot finalize) or not (we can/should finalize).