ENG-19635: Do not delete segments on transient close
Transient cursors are not used to detect when segements should be
deleted. However they could still cause deletion when they were closed
since the cursor close logic would still delete segments even if the
cursor which was being closed was transient.

If the cursor is transient do not delete any segments.