Optimize away some deep copier/finalize operations, etc. (#7297)
* [Careless] optimize away some deep copier/finalize operations

on the new/different memory model:

1. Previously, both src/dst of moved tuples in batch removal are
captured in hook memory. Using the new memory model, we can safely only
capture the dst tuple (that gets overwritten) in hook.

2. Previously, inside the callback of batch removal, index is always
updated for the movement. Now, since we are guaranteed that source tuple
(the memory that gets copied to somewhere else) remains visible and
intact only in snapshot region, we *replace* index on the two tuples
only when the table is not frozen; when it is frozen, we simply add the
dst tuple to index, therefore keeping the source tuple valid in the
index. This part is most suspicious to me: should I add index to the
tuple that hook shallow copies (see #3)?

3. Previously, when adding hook memory entry for deletes or updates, we
always use deep copier when applicable. Now, we only invoke deep copier
on update: we use shallow memcpy for deleted tuples, and let batch
removal call back discriminate the frozen case, when it does not
finalize (i.e. call `decreaseStringMemCount()' and
`freeObjectColumns()') if frozen.

* Fix: in batch removal, we can

no longer safely finalize all tuples that gets removed (i.e. those not being
compacted to holes) at frozen time, since we spared copy (and deep copy)
into hook memory. To correctly do this, we need to differentiate the
removed tuples, checking if it lies inside frozen boundary (in which case,
we cannot finalize) or not (we can/should finalize).

* Added single removal API

as light weight alternative to batch removal API when only
used to remove a single tuple

* Prepare to use single-removal API in persistenttable

* Separate add() into 2 methods

for delete and update.

* Fixed failed CoveringCellIndexTest

when using the light weight single tuple removal operation.

* Addressed Nate PR comments