ENG-19814: Add force response node
Add a force response node to the SSM so that all members of the state
machine post a result for the latest version. The force response node is
only created when a host which is initializing does not see a response
from all other members. This is to resolve a race with a proposer going
down and a new member initializing at the same time and the initializer
cannot determine if a proposal is being processed or not.