ENG-19869: Increase next available ticket precision
When the rate limit is greater than 1,000,000 any request for a ticket
value less than 1/1,000,000 the limit will not be counted towards the
limit. The issue comes from the fact that the limit is specified in
tickets per second and time is tracked in microseconds. Since the next
available ticket in micro time was tracked as a long any ticket request
of less than 1/1,000,000 the rate would be rounded down to 0.

By changing the next available ticket to a double any requests which are
below the microsecond precission can still be tracked and count towards
the limit.