ENG-19984: Add updateEntries API to BinaryDeque
Add a new visitor style API which iterates over all entries in reverse
order allowing each entry to be updated independentaly. If there are no
changes to a segment it is untouched. However, if either an entry is
deleted or updated the segment will be recreated by writing a new file
and renaming it to replace the existing segment.