ENG-19985: EntryUpdater notify on end of segment and updates
Update EntryUpdater to extend AutoCloseable so that the updater can be
notified when all updates are completed and any resources being held can
be released. Also, add segmentComplete() to notify the updator when a
segment has completed processing and all buffers to write to the segment
are no longer in use.