ENG-19985: Allow adding new entries during update
When the end of a segment is reached EntryUpdater.update is still
invoked with a null entry and the implementation at this ponit can
choose to add a new entry in the segment or keep the current end point.

This is needed for topic compaction which needs to leave behind
tombostones when entries are deleted and it is more efficient to merge
adjacent deleted entries into a single tombstone.