ENG-20070: Handle null return by read
CSVTableSaveFile.read() is documented to return null when the end of the
data is reached however CSVTableSaveFile.convertTableSaveFile() assumed
that it never did. Handle when read() returns null as end of data and
break out of the read loop.

Update SnapshotConverter to print full exception stack traces when
exceptions are encountered, so that future debugging is easier.