ENG-19933, avoid using getNValue() directly in TableTuple.debug(). (#130)
debug() is usually called when something is broken, to collect useful information for further diagnosing. If
the tuple was corrupted, calling getNValue() would trigger a new SQL error. As a result, the collateral exception
masks the true case of error. So in case of handling corrupted tuple data, debug() will at least display the
original error.

Co-authored-by: Lukai Liu <lliu@voltdb.com>