ENG-20395: Handle all byte buffers as remnant
Correctly handle the remnant in SocketJoiner.processSSC. Update the
handling so both heap and direct buffers can be used. Always read the
size of the string from the buffer so that the string can be decoded
when asserts are disabled.

VoltProjectBuilder: Add method to be able to enable internal SSL.