StreamBlock.h: Rollback SpTxnId too
The export tuple stream does not use lastSpUniqueId the same way DR does
and instead sets this value on every append tuple which means that if a
transaction is rolled back this value needs to be rolled back too.

Update ExportStreamBlock::truncateExportTo also rollback the spTxnId too
so that it correctly corresponds to the last spTxnId which is in the
block.