ENG-20407: Add TopicTupleStream
Create a new class TopicTupleStream which extends from ExportTupleStream
so that it can be used in StreamedTables instead of ExportTupleStream.
This required some methods to be made virtual as well as a new virtual
method to allocate the block which backs the stream.

TopicTupleStream writes tuples in kafka record format to the block using
the encoders for the key and value.