SSL: Cleanup cipher selection
Move excluded cipher list to SSLConfiguration to be with other sets.

Remove current preferred cipher list since all of the ciphers which are
in this set which are not in the GCM set are in the exclude set so they
cannot be used.

Rename GCM set to preferred set and add all TLSv1.3 ciphers to the set.

Create a common caching CipherSuiteFilter to be used by SslContexts so
that all SSLEngines will have the same cipher list.