SaveRestoreBase: Do not delete voltdbroot between tests
SaveRestoreBase extends from RegressionSuite which starts a single
cluster and reuses the same cluster for every test. However,
SaveRestoreBase deletes the current working voltdbroot directory before
and after each test. This means that after the first test finishes the
cluster which is running for all of the other tests is running out of a
deleted directory. This just happens to work now as long as no hosts are
restarted.

TestStatisticsSuite.testStatsDoesNotTimeoutDuringRejoin does restart a
host which can cause problems since if that is not the first test to run
the restart will fail and all other tests after it could have errors
too.

Instead of deleting the directories between every test only delete the
directories if a complete shutdown is needed after this test.