ENG-20548: Calculate a set of replicable tables
The hash of a tables signature is not reliable to determine if the local
table is the same as the remote. In order to do that the
DRCatalogDiffEngine will calculate a list of tables that are compatible
on both the producer and consuer. This will be done at three points.
Initial connection to the remote side, local catalog update and remote
catalog update. The list must be generated in an MP procedures so that
it is performed transactionally.

Once the list is generated it will be sent down to the ee to enforce
which tables can be the target for replication from which source
clusters.