ENG-20548: Add producer catolog commands to snapshot
When performing command log recovery the replicable tables need to be
set before processing any apply binary logs. If they are not set then
all logs will be rejected.

Add the producer catalog commands into the snapshot and restore the
commands and the replicable tables during SnapshotRestore. Both need to
be restored so that if there is a local catalog update in the command
log then the replicable tables after that point are updated.

This moves the catalog command storage out of ConsumerDRGatewayImpl
because that is not available during command log replay.