ENG-20714: Add a way to wrap the FileChannel for testing
Add a new static in NativeSnapshotDataTarget which allows tests to
modify the FileChannel created by a snapshot target. Once the modifier
is used by a snapshot it is immediatly reset back to a no-op so that it
does not bleed into other snapshots.

Add a utility class for injecting specific types of errors or problems
into the snapshot targets.