ENG-21356, add error and debug logging to interface address discovery (#650)
This adds some logging to interface discovery in RealVoltDB, in the hope that it'll shed some light on an address-resolution issue that has so far proven to be refractory.

From the point of view of the reported symptoms, the most likely thing that happened was that we found no address so (uselessly) decided to use 'localhost' instead. I have not changed that, but at least there's a warning if we do so.

I found a theoretical bug: if an interface had more than one IPv6 address, and a link-local address was seen after a real address, we'd forget the real address. This is now fixed. I don't know whether this was the customer's problem, but they were using IPv6 only.