ENG-20828, command specific log4j default (#939)
Because of the 'undifferentiated pile of code' nature of VoltDB, any 'voltdb' command that runs Java code might include a VoltLogger instance somewhere. For voltdb init and voltdb start, which run the VoltDB server. we want the real log4j.xml, because those commands use the appropriate voltdbroot/log directory. For voltdb get, even though this runs the VoltDB server jar, it does not (and should not) write to voltdbroot/log, so we want the utility-log4j.xml file to get console output. Similarly for collect, which is a separate jar, but using common source code, For mask, there's not currently any logging, but let's future-proof it.