Eng 21974 report file descriptor limit and usage (#1027)
Added new `@Statistics` about file descriptor and client connections limits and usage. Under `LIVECLIENTS_CONNECTIONS` there will be a new table describing what is the limit of file descriptor for VoltDB process, what is the usage of file descriptors as well as client connections limit and number of client connections.

Because of how grouping of tables under selectors work I had to add a new selector `LIMITS`. It is possible to query it individually to get just that table.

The limits and usage are refreshed every 10 minutes, and at that moment warning is printed if limit is too low or usage is too hight. Exception is the number of connections which is always up to date.

Note 1: maximum number of connections is equal to descriptor limit minus 300. It is unchanged behaviour from before this PR. 
Note 2: Warning about high usage of descriptors is printed when we have less than 300 descriptors left.