remove ttl type check (#1244)
These days, only TIMESTAMP is accepted for a TTL column type. This is checked in DDLCompiler.

The HSQL parser also checked type but it had a different set of permitted types.

The solution, possibly counter-intuitive, is to make the HSQL parser accept an even wider range of types, by removing the check altogether. That way the error is always detected in DDLCompiler, and we get a consistent error message for (say) strings and integers.