ENG-22390, remove dependency of VoltLogger on RateLimitedLogger (#1257)
Tracking of rate-limiting per message used a cache of RateLimitedLogger objects. This is overkill since the only thing that was actually needed from the RateLimiterLogger was the single 'time we last logged' longword. We can reduce overhead, and untangle a two-way dependency, by having VoltLggger keep a simpler cache of time-last-logged.

Actual changes:

- New LogRateLimiter class for use of VoltLogger (a separate class just for cleanliness) which only answers the question "should we log this now?"

- LogRateLimiter uses a much simpler cache structure

- VoltLogger is independent of the old RateLimitedLogger

- RateLimitedLogger.tryLogForMessage is deprecated

No change needed to callers that are using voltLogger.rateLimitedLog() etc.