always return call stats even if there are none (#1311)
Procedure stats and compound proc call stats are registered per proc.

If there were no compound procs in the system, a stats query for COMPOUNDPROCCALLS would result in an error message rather than an empty table. Though this was basically a factor of the framework, compound procs is the only observed incidence (there can never be no procedures).

There's little scope for fixing the framework since the framwrok doesn't know how to make an empty table for a particular stats selector. So, kludgery. We always register one dummy procedure. Fortuitously, this generates no call stats, since it makes no calls.

The dummy proc is registered at startup and again when the catalog is changed (this last parallels pre-existing treatment of procedures; we don't hold on to the stats if procs might have changed).