count a timed invocation after other counts updated (#1318)
he reported problem is that sometimes @statistics can see procedure stats where the min execution time = max long int and max exec time = min long int.

Those numbers are the (quite reasonable) initialized states of the stats. You want the min observed value to be initialized to a big number, so the first observed actual value is guaranteed to be less, without special-case code.

The fundamental issue is the reading the stats is not synchronized with updating the stats, so from time to time you may see intermediate updates. I do not care to add explicit sync since that may cause @statistics to delay transaction execution.

There's already code to skip stats if the timed-transaction sample count is zero (i.e., there are no times recorded). But we increment that before recording the time, creating a window where count is positive but the times have not been adjusted.

I think we can make this better by a simple reordering. The stats-row iterator skips entries with a zero timed-transaction count. (It's too late if we actually get to updateStatsRow, we have to report something, and reporting (say) max = 0 rather than max = a very big number would be very wrong).

This of course guarantees nothing under language rules. (Come to that, the lack of sync allows for the result that the stats agent thread will never ever see any updates make by the site thread; but we seem to be living with that).