Eng 22414 periodic logging (#1327)
This adds a little information to the daily logging - version and build string, local node state and uptime, cluster safety, kubernetes yes/no.

It also unifies startup logging and daily logging.

The idea of this is to get useful changes near the top of every log file. (It'd be nice if this were actually sync'd with log rollover, but that's a task for another day. Here we just "hope" it happens soon after log rollover).

I moved all of this logging stuff to a separate class since RealVoltDB has become a bit of a dumping ground. In doing that. I found that a couple of "logging" routines actually made config updates. I have not changed that. But I did comment it.