make DailyLogTask scheduling work with DailyMaxRollingFileAppender (#1333)
We like to put status/config info in every log file. The code to do that attempts to run 30 secs after log rollover. It got broken when we switched from DailyRollingFileAppender to DailyMaxRollingFileAppender.

This change fixes the immediate problem and  also moves the bulk of the logic from RealVoltDB to VoltLog4jLogger -- it's specific to log4j and therefore should be kept out of RealVoltDB.